refactor: move torch functions to root and mnist eval test
