+++ wildfire/evaluate.py
@@ -2,10 +2,8 @@ import torch
 from functools import wraps
 
 from wildfire.functional import structure_map
-from wildfire.torch import module_device, module_train
-from wildfire.torch import (
-    to_device, step
-)
+from wildfire import module_device, module_eval
+from wildfire.to_device import to_device
 
 
 def cpu_detach(x):
@@ -15,21 +13,20 @@ def cpu_detach(x):
         return x
 
 
-def train(model, optimizer, n_batches_per_step=1):
+def evaluate(model):
     device = module_device(model)
 
     def decorator(process_batch):
 
         @wraps(process_batch)
         @to_device(device)
-        @step(optimizer, n_batches_per_step=n_batches_per_step)
+        @torch.no_grad()
         def _process_batch(*args, **kwargs):
-            with module_train(model):
+            with module_eval(model):
                 return structure_map(
                     cpu_detach,
                     process_batch(*args, **kwargs),
                 )
-
         return _process_batch
 
     return decorator
