+++ wildfire/numpy/figure_to_numpy.py
