+++ wildfire/numpy/numpy_seed.py
