+++ wildfire/train.py
@@ -2,8 +2,10 @@ import torch
 from functools import wraps
 
 from wildfire.functional import structure_map
-from wildfire.torch import module_device, module_eval
-from wildfire.torch.to_device import to_device
+from wildfire import module_device, module_train
+from wildfire import (
+    to_device, step
+)
 
 
 def cpu_detach(x):
@@ -13,20 +15,21 @@ def cpu_detach(x):
         return x
 
 
-def evaluate(model):
+def train(model, optimizer, n_batches_per_step=1):
     device = module_device(model)
 
     def decorator(process_batch):
 
         @wraps(process_batch)
         @to_device(device)
-        @torch.no_grad()
+        @step(optimizer, n_batches_per_step=n_batches_per_step)
         def _process_batch(*args, **kwargs):
-            with module_eval(model):
+            with module_train(model):
                 return structure_map(
                     cpu_detach,
                     process_batch(*args, **kwargs),
                 )
+
         return _process_batch
 
     return decorator
