+++ wildfire/update_cpu_model.py
